packages/backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
python-multipart
moviepy
supabase
python-jose[cryptography]
slowapi